--- tests/propose.py
@@ -1,46 +1,0 @@
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the Apache 2.0 License.
-import sys
-import os
-import infra.proc
-
-import e2e_args
-import getpass
-import os
-import time
-import logging
-import multiprocessing
-from random import seed
-import infra.ccf
-import infra.proc
-import infra.jsonrpc
-import json
-
-from loguru import logger as LOG
-
-
-def run(args):
-    hosts = ["localhost", "localhost"]
-
-    with infra.ccf.network(
-        hosts, args.build_dir, args.debug_nodes, args.perf_nodes, pdb=args.pdb
-    ) as network:
-        primary, others = network.start_and_join(args)
-
-        for i in range(1000):
-            result = network.propose(1, primary, "add_user", "--user_cert=member3_cert.pem")
-            assert not result[1]["completed"]
-
-
-if __name__ == "__main__":
-
-    def add(parser):
-        parser.add_argument(
-            "-p",
-            "--package",
-            help="The enclave package to load (e.g., libloggingenc)",
-            default="libloggingenc",
-        )
-
-    args = e2e_args.cli_args(add)
-    run(args)
