+++ tests/requirements.txt
@@ -2,3 +2,4 @@ paramiko
 msgpack
 loguru
 coincurve
+psutil
